requirements.txt
async-timeout==4.0.3
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-extended==0.2
Flask-JWT-Extended==4.6.0
flask-redis==0.4.0
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.4
SQLAlchemy==2.0.30
typing_extensions==4.12.0
uuid==1.30
Werkzeug==3.0.3
WTForms==3.1.2
